refactor: move GUI to a separate ui package
