+++ src/main/java/duke/command/AddCommand.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
