+++ src/main/java/duke/command/ExitCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import duke.Storage;
 import duke.TaskList;
-import duke.Ui;
+import duke.ui.Ui;
 
 /**
  * A command to exit the program.
